Data validation Changes in cost type Changed details update

+++ src/main/java/shopApp/model/product/Product.java
@@ -14,5 +14,5 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String productName;
-    private Double productPrice;
+    private Integer productPrice;
 }
